Added combined plot with project
The approach basically plot the same data with
different colour mapping. The alpha is set based on the fraction of
the projected spectral function compared to that of the total.
However, the colour mapping should start from white while the existing
ones start from some very fint colour instead....
This approach allows any number of projected spectral functions to
be used, instead of limited to three species compared to the RGBA
approach.