Allow custom colour map to be used
The colour map create is from white to the terminal colour. If a
hex colour code is passed then it is used as the terminal colour.
If an existing colour map is used then the terminal colour of that
colour map is used instead.

Ensure the 1-based indexing is used when passing atomic indices.
Allow "-" to be used for specifying ranges.