Add functionality for getting atomic projections for correct spectral function plotting (#4)
* Add functionality for getting atomic projections for correct spectral function plotting

* Added procar reading code

* Added weighting by atoms/orbitals

Allows the weighting of each band to be constructed and used for
making spectral functions.
The the spectral weights are multiplied by the weights, if avalaible
during the construction of the spectral function.

* Add CLI plotting interface for atomic weighting

* Added code for plotting multiple projections

Create a figure with subplots showing different projection channels

* Minor tweak in the wording

* Mintor tweak

* Added documentation system

* README md update

* Update documentation build instructions

* Update the README.md

* Added combined plot with project

The approach basically plot the same data with
different colour mapping. The alpha is set based on the fraction of
the projected spectral function compared to that of the total.
However, the colour mapping should start from white while the existing
ones start from some very fint colour instead....
This approach allows any number of projected spectral functions to
be used, instead of limited to three species compared to the RGBA
approach.

* Allow custom colour map to be used

The colour map create is from white to the terminal colour. If a
hex colour code is passed then it is used as the terminal colour.
If an existing colour map is used then the terminal colour of that
colour map is used instead.

Ensure the 1-based indexing is used when passing atomic indices.
Allow "-" to be used for specifying ranges.

* Added MgO example in the documentation

* Minor change

* Fix test

* Change placeholder file for theory page

* Minor link update

* Parital commit of theory section

* Trimmed down the code base

Remove modules and code that are not needed.

* Added effective mass calculation, doc update

Effective mass can now be calculated based the effective band structure.
Note that only the apparent band-edge cases can be treated, and
it can fail for cases where there are degenerate bands at the CBM/VBM.
The results should be cross-checked with the effective band structure
carefully.

* Add band edge detect tolerance

Effective mass may not work for spin and soc cases due to issue
with band edge detection based on CBM/VBM values. Most robust
ways should be introudced. Probably the number of electrons should
be recorded in JSON file to facilitate the detection....

* Add test data

* Improve the robustness of band extrema detection

Assuming the extreme can only occur at high symmetry points

* Fix tests

* bump up the version number

* Fix tests

* Updated the instructions for installation

* Updated project.toml

* Added plotting module

New functionalities:
- Allow effective mass raw data to be plotted
- Updated the routine of effective mass extrema detection
- New plotting module

Bug fix:
- Fixed a bug of incorrect reciprocal space coodinate conversion.
Previously the cartesian coordinates can be wrong, leading to
incorrect effective masses if the cell matrix is non-diagonal.

* Fix tests

Improved effective mass printing

* Fixed the labelling when printing the symmetry information for the primitive and supercell

* Fix CLI test

* Update band extrema and degeneracy detection

Allow effective mass fitting results to be plotted.

* Use `tabulate` package for displaying tables

* Fix tests

* Improved projection plotting

* Allow any color to be used for combined projection map

Allow colored atomic-projection by interpolating from any terminal
colors within a color space. This may still result in "white" colours
if there are too many atomic elements. Hence when constructing the
combined plots care should be taken. This would not be a problem if
the orbitals are well separated along the energy axes, however.

* Added coverage report in the CI workflow

* CI Update

* CI and Readme update

* Remove duplicates from help options in `cli.py`, and show defaults for some important parameters (so users know how to change them / what they'll be changing them relative to)

* Fix typos

* Add "Energy (eV)" to y-axis for projected plots

* Update to handle multiple-digit numbers in `atoms_idx` (particularly useful for supercells with many atoms), so e.g. `1-20|21-40|41-80` can be parsed (previously threw a `ValueError`)

* Add NaBiS2 example docs page (also shows all these small update commits in action 👆)

* Make docs badge in `README.md` link to docs

* Add NaBiS2 example to sidebar in docs

* Docs update; add CLI and Python API sections

---------

Co-authored-by: Bonan Zhu <zhubonan@outlook.com>
Co-authored-by: adair-nicolson <56972784+adair-nicolson@users.noreply.github.com>