Using time reversal symmetry for SC Kpoints
Time reversal symmetry can be used to reduce the number of kpoints
in the supercell. This because kpoints that are NOT related by
time reversal symmetry may have the related K in the supercell
related by time reversal.

This reduces the number of kpoints sampled in the supercell. Backward
compatility should be maintained as the change only affects the number
of SC kpoints generated initially.

Kpoints in the wavefunction are checked for equivalent taking account
of time reversal symmetry, if applicable.